requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
apscheduler==3.10.4
    # via python-telegram-bot
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
distro==1.9.0
    # via openai
emoji==2.12.1
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   openai
    #   python-telegram-bot
idna==3.7
    # via
    #   anyio
    #   httpx
jiter==0.5.0
    # via openai
loguru==0.7.2
    # via
    #   -r requirements.in
    #   sentry-sdk
openai==1.40.2
    # via -r requirements.in
pydantic==2.8.2
    # via openai
pydantic-core==2.20.1
    # via pydantic
python-telegram-bot==21.3
    # via -r requirements.in
pytz==2024.1
    # via
    #   apscheduler
    #   python-telegram-bot
sentry-sdk==2.12.0
    # via -r requirements.in
six==1.16.0
    # via apscheduler
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   emoji
    #   openai
    #   pydantic
    #   pydantic-core
tzlocal==5.2
    # via apscheduler
urllib3==2.2.2
    # via sentry-sdk
